Split up test util files
